+++ vscode/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "rl" extension will be documented in this file.
 
+## [v0.3.1]
+
+- adds all length and time SI units
+- warn on unused variables
+
 ## [v0.2.0]
 
 - adds micro and nano units